backend/requirements.txt
# FastAPI
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Pydantic
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Firebase Admin
firebase-admin==6.4.0

# Supabase
supabase==2.3.4

# Google Gemini AI
google-generativeai==0.3.2

# Document Processing
PyPDF2==3.0.1

# Database
psycopg2-binary==2.9.9

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

PyPDF2
supabase